Seeds are 64-bit: every seed from 0 to 2**64 - 1 is accepted, and every seed below 2**32 is the market it was
The engine took a u32 seed and refused anything from 2**32 up, so a sealed
battery's hidden seed was one of 2**32 per roster and could be found by
simulating every one against a market's first prices.

The raw generator adds the whole u64 seed to the PCG state, which is the
same arithmetic as before for a seed below 2**32. The stream derivation lets
the high 32 bits in only when they are non-zero: a narrow root takes the old
formula unchanged, and a wide one starts each stream from all 64 bits of
splitmix64(s ^ splitmix64("SD64" << 32 | k)) on sequence 768 + k, a
bijection per stream on a sequence no narrow root uses. Surgery seeds widen
the same way (sequence 1024 + k). rust/src/rng.rs and docs/MODEL.md state
the formulas.

Every digest is unchanged: sim 4528d93a, sha256 f4a81e0b, meta 8804ef0e,
bonds b8798418, book 1e7f1884, the 19 preset rows (combined f131be87) and the
wasm probe 1c5acabf. A new row, tests/known_answer_seed64.py, hashes seed
2**63 + 12345 on pt-v19 (cef62229) and is the sixth digest the determinism
workflow compares.

Every surface takes the wide range and refuses a negative seed, one of
2**64 or more, a float or a bool with a ValidationError naming the range:
Engine, EngineBatch, Universe.random, GameRng, evaluate, reference agents,
World and its surgery seeds, the gym, run_many, sweep, rank, commit, reveal,
sealed_battery, Checkpoint, RunManifest, manifest.verify (which masked its
seed to 32 bits), StrategySpec, edgar, atlas and the MCP tools. WebAssembly
takes a Number up to MAX_SAFE_INTEGER or a BigInt, checked in node against
the native digest.

+++ rust/src/engine.rs
@@ -618,7 +618,7 @@ impl Engine {
     /// changes. A 30-name universe and a 100-name universe from one seed have
     /// nothing to do with each other.
     pub fn new(
-        seed: u32,
+        seed: u64,
         companies: Vec<TickCompany>,
         economy: EconomyState,
         central_bank: CentralBankState,
@@ -983,7 +983,7 @@ impl Engine {
     /// preset-constructed engine reproduces the const build's trajectories
     /// bit for bit, draw for draw — the phase-1 acceptance gate.
     pub fn with_params(
-        seed: u32,
+        seed: u64,
         companies: Vec<TickCompany>,
         economy: EconomyState,
         central_bank: CentralBankState,
@@ -1018,7 +1018,7 @@ impl Engine {
     /// is what a caller naming an opening asked for, and it is what every
     /// preset before pt-v18 did.
     pub fn with_params_from_opening(
-        seed: u32,
+        seed: u64,
         companies: Vec<TickCompany>,
         economy: EconomyState,
         central_bank: CentralBankState,
@@ -6734,7 +6734,7 @@ mod tests {
         }
     }
 
-    fn engine(seed: u32) -> Engine {
+    fn engine(seed: u64) -> Engine {
         Engine::new(
             seed,
             vec![company("A", 100.0), company("B", 50.0), company("C", 220.0)],
@@ -6748,7 +6748,7 @@ mod tests {
     /// agent's fills reach the first tick's `order_volumes`. pt-v20, the
     /// default from 0.8.5, turns `fill_impact_coefficient` on and routes them
     /// through the agent-facing book's pending flow instead.
-    fn engine_v19(seed: u32) -> Engine {
+    fn engine_v19(seed: u64) -> Engine {
         Engine::with_params(
             seed,
             vec![company("A", 100.0), company("B", 50.0), company("C", 220.0)],
@@ -6813,7 +6813,7 @@ mod tests {
             }
             e.prices()
         };
-        let seeds: Vec<Vec<f64>> = (1u32..=4).map(run).collect();
+        let seeds: Vec<Vec<f64>> = (1u64..=4).map(run).collect();
         for (i, a) in seeds.iter().enumerate() {
             for b in seeds.iter().skip(i + 1) {
                 assert!(
@@ -8079,8 +8079,8 @@ mod tests {
 /// Returns `None` for an unknown preset.
 pub fn fixed_simulation_digest(
     size: usize,
-    universe_seed: u32,
-    seed: u32,
+    universe_seed: u64,
+    seed: u64,
     days: usize,
     ticks: usize,
     preset: &str,
